requirements.txt
gradio==3.1.0
pandas==1.4.3
python-dotenv==0.20.0
langchain==0.1.17
langchain_google_genai==0.0.7
httpx==0.23.0
httpcore==0.15.0
